+++ src/main.tsx
@@ -1,30 +1,13 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import Index from './index.tsx'
-import LoginForm from './login.tsx'
-import './main.css'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App.tsx";
+import { Provider } from "react-redux";
 import { store } from "./store";
-import { Provider } from 'react-redux';
 
-import {createBrowserRouter, RouterProvider,} from "react-router-dom";
-import "./main.css";
-
-const router = createBrowserRouter([
-    {
-        path: "/",
-        element: <Index/>,
-    },
-    {
-        path: "/login",
-        element: <LoginForm/>,
-    },
-]);
-
-ReactDOM.createRoot(document.getElementById('root')!).render(
-    <React.StrictMode>
-        <Provider store={store}>
-            <RouterProvider router={router}/>
-        </Provider>
-    </React.StrictMode>,
-)
-
+ReactDOM.createRoot(document.getElementById("root")!).render(
+	<React.StrictMode>
+		<Provider store={store}>
+			<App />
+		</Provider>
+	</React.StrictMode>
+);