+++ src/register.tsx
@@ -3,41 +3,99 @@ import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
 import {library} from '@fortawesome/fontawesome-svg-core'
 import {fab} from '@fortawesome/free-brands-svg-icons'
 // import { selectUser } from "./store";
+import { auth } from "./firebase";
+import { createUserWithEmailAndPassword } from "firebase/auth";
+import {FormEvent, useState} from "react";
+import {useDispatch} from "react-redux";
+import {login} from "@/store/auth/authSlice.ts";
+import {useNavigate} from "react-router-dom";
 
 library.add(fab)
 
-
-
 function SignUpForm() {
+
+    const dispatch = useDispatch();
+    const navigate = useNavigate();
+
+    const [firstName, setFirstName] = useState("");
+    const [lastName, setLastName] = useState("");
+    const [username, setUsername] = useState("");
+
+    
+    function handleFirstNameChange(e: any) {
+        setFirstName(e.target.value);
+    }
+    
+    function handleLastNameChange(e: any) {
+        setLastName(e.target.value);
+    }
+    
+    function handleUsernameChange(e: any) {
+        setUsername(e.target.value);
+    }
+
+    function handleSignUpWithEmailAndPassword(event: FormEvent) {
+        event.preventDefault();
+
+        // @ts-expect-error annoying
+        const email = event.target.email.value;
+        // @ts-expect-error annoying nagging bitch
+        const password = event.target.password.value;
+
+        // check if email is valid, password is valid
+
+        createUserWithEmailAndPassword(auth, email, password)
+            .then((userCredential) => {
+                // Signed up
+                console.log("Signed up successful");
+                const user = userCredential.user;
+                dispatch(login(user));
+                navigate("/");
+            })
+            .catch((error) => {
+                const errorMessage = error.message;
+                console.error(errorMessage);
+            });
+    }
+
     return (
         <div className="lg:flex h-screen bg-primary">
             <div className="w-full bg-white flex items-start justify-start p-4 lg:hidden">
-                <img src='/Logo.png' alt="Forklore logo" className="h-10 mx-auto"/>
+                <img src='/logo.png' alt="Forklore logo" className="h-10 mx-auto"/>
             </div>
             <div className="hidden lg:block lg:w-1/2 bg-secondary flex items-start justify-start p-4">
-                <img src='/Logo.png' alt="Forklore logo" className="hidden lg:block h-10"/>
+                <img src='/logo.png' alt="Forklore logo" className="hidden lg:block h-10"/>
             </div>
-            <div className="w-full lg:w-1/2 bg-white flex items-center justify-center">
-                <div className="flex items-center justify-center h-screen">
-                    <form className="flex flex-col w-4/5 max-w-lg mx-auto items-center">
+            <div className="w-full lg:w-1/2 bg-white items-center">
+                <div className="flex flex-col items-center justify-center h-screen">
+                    <form onSubmit={handleSignUpWithEmailAndPassword} className="flex flex-col w-3/5 max-w-lg mx-auto items-center">
                         <h2 className="mb-4 text-2xl text-primary">Sign up</h2>
                         <div className="flex flex-row items-stretch justify-center w-full">
-                            <input className="mb-4 mr-1 w-full p-2 border border-primary rounded-full" type="text"
-                                   placeholder="First Name" required/>
-                            <input className="mb-4 ml-1 w-full p-2 border border-primary rounded-full" type="text"
-                                   placeholder="Last Name" required/>
+                            {/* First name field*/}
+                            <input value={firstName} name={"firstName"} className="mb-4 mr-1 px-4 py-2 w-full p-2 border border-primary rounded-full"
+                                   type="text"
+                                   placeholder="First Name" onChange={handleFirstNameChange} />
+                            {/* Last Name Field*/}
+                            <input value={lastName} name={"lastName"} className="mb-4 ml-1 px-4 py-2 w-full p-2 border border-primary rounded-full"
+                                   type="text"
+                                   placeholder="Last Name" onChange={handleLastNameChange} />
                         </div>
-                        <input className="mb-4 w-full p-2 border border-primary rounded-full" type="text"
-                               placeholder="Username" required/>
-                        <input className="mb-4 w-full p-2 border border-primary rounded-full" type="password"
+                        <input value={username} name={"username"} className="mb-4 w-full px-4 py-2 border border-primary rounded-full" type="text"
+                               placeholder="Username" onChange={handleUsernameChange} required/>
+                        <input name={"email"} className="mb-4 w-full px-4 py-2 border border-primary rounded-full" type="text"
+                               placeholder="Email" required/>
+                        <input name={"password"} className="mb-4 w-full px-4 py-2 border border-primary rounded-full" type="password"
                                placeholder="Password" required/>
                         <button className="w-1/2 p-2 bg-primary text-white rounded-full hover:bg-accent"
-                                type="submit">Submit
+                        > Sign up
                         </button>
-                        <div className="flex items-center w-full mt-4">
-                            <hr className="flex-grow border-t border-secondary"/>
+                    </form>
+                    <a className="mt-4 text-primary hover:text-accent hover:underline" href="/login">Already have an
+                        account? Sign in</a>
+                    <div className="flex justify-center items-center w-full mt-4">
+                            <hr className="w-1/4 border-t border-secondary"/>
                             <p className="mx-4 text-secondary">OR</p>
-                            <hr className="flex-grow border-t border-secondary"/>
+                            <hr className="w-1/4 border-t border-secondary"/>
                         </div>
                         <p className="mt-4 text-primary">Sign Up with:</p>
                         <div className="flex w-full justify-center mt-4">
@@ -57,7 +115,6 @@ function SignUpForm() {
                                 <FontAwesomeIcon icon={['fab', 'apple']} size="2x"/>
                             </a>
                         </div>
-                    </form>
                 </div>
             </div>
         </div>