+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 # Project change log
 ===================
 
+### 0.0.8 2024-17-05
+* Improved design of register and sign in page
+* Added basic design of home page that includes recipes
+* Fixed issues with the website not working correctly on firebase hosting
+* Restructured frontend code for code scalability
+* Addes sample recipe data (for testing)
+
+### 0.0.8 2024-05-06
+* Added profile button on navbar when user is signed in
+* Added (working) sign out button in navbar profile menu
+* Made it so every page except the register and login will show the header and footer.
+
+### 0.0.7 2024-05-03
+* Added `/recipe` route template.
+
+### 0.06 2024-05-02
+* Added animation on scroll for navigation bar
+* Added footer
+* Added 404 page
+
 ### 0.0.5 2024-05-01
 * Added navigation bar (for desktop view)